Change coordiantes to float (from int)

+++ src/main/java/org/demo/rsotf/model/Area.java
@@ -2,9 +2,9 @@ package org.demo.rsotf.model;
 
 public class Area {
 
-    private int x0, y0, x1, y1;
+    private float x0, y0, x1, y1;
 
-    public Area(int x0, int y0, int x1, int y1) {
+    public Area(float x0, float y0, float x1, float y1) {
         if (x0 <= x1) {
             this.x0 = x0; this.x1 = x1;
         } else {
@@ -18,7 +18,7 @@ public class Area {
         }
     }
 
-    public boolean contains(int x,  int y) {
+    public boolean contains(float x, float y) {
         return x0 <= x && x <= x1 && y0 <= y && y <= y1 ;
     }
 }
